sentimental processing added commit

+++ src/main/java/twitter/opinion/mining/response/pojo/AnalysedData.java
@@ -2,22 +2,22 @@ package twitter.opinion.mining.response.pojo;
 
 public class AnalysedData {
 
-	public AnalysedData( String name,Integer score) {
+	public AnalysedData( String name,Double score) {
 		this.name=name;
 		this.score=score;
 	}
 	private String name;
-	private Integer score;
+	private Double score;
 	public String getName() {
 		return name;
 	}
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getScore() {
+	public Double getScore() {
 		return score;
 	}
-	public void setScore(int score) {
+	public void setScore(Double score) {
 		this.score = score;
 	}
